+++ my-app/src/App.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import './App.scss';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
-import MovieInfo from './container_components/Movie_info';
-import SearchPage from './container_components/Search';
+import MovieInfo from './container_components/Movie_info_container';
+import SearchPage from './container_components/Search_container';
 import NotFound from './presentational_components/NotFound';
 import ErrorBoundary from './shared_components/ErrorBoundry/errorBoundry';
